update :: 함수 클래스로 분할
